+++ language/dutch_utf-8.php
@@ -356,6 +356,9 @@ You may either move existing events to a new calendar, or delete those events.',
 'enable_comments' => 'Enable Comments?',
 'closed'        => 'Closed',
 'event'         => 'Event',
+'timezone'      => 'Timezone',
+'tz_local'      => 'Guest&apos;s local timezone',
+'tz_select'     => 'Select Timezone',
 );
 
 // Localization of the Admin Configuration UI