+++ public_html/ajax.php
@@ -22,9 +22,9 @@ case 'getloc':
     $B = array(
         'id'        => '',
         'title'     => '',
-        'street'    => '',
+        'address'    => '',
         'city'      => '',
-        'province'  => '',
+        'state'  => '',
         'country'   => '',
         'postal'    => '',
         'lat'       => '',
@@ -36,7 +36,7 @@ case 'getloc':
     }
     $id = isset($_GET['id']) && !empty($_GET['id']) ?
                     COM_sanitizeID($_GET['id']) : '';
-    $status = LGLIB_invokeService('getInfo', 'locator',
+    $status = LGLIB_invokeService('locator', 'getInfo',
             array('id' => $id), $A, $svc_msg);
     if ($status == PLG_RET_OK) {
         if (!$A) {