+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.4.2
 Released 2017-01-21
 - Fix calls to commentsupport function in another plugin
+- Add timezone support for events
 
 ## v1.4.1
 Released 2017-01-21